prep for retooling
This is a series of patches to migrate TamTamSynthLab to the 0.86 toolbars. The first two patches are meant to consolidate the existing toolbar code. The third patch creates a separate toolbar for the record buttons (currently disabled) so that the duration slider and trash can be moved to the main toolbar. The fourth patch adds support for the new toolbars.